설문조사 기능 구현후 충돌 병합

+++ Ironman/src/main/java/com/Ironman/back/dto/SurveyDto.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class SurveyDto {
-    private Integer height;
-    private Integer weight;
+    private float height;
+    private float weight;
     private String activityLevel;
     private String pushUp;
     private String plank;
